# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Backend cho ứng dụng SmartSchedule - Lịch hẹn thông minh",
        "title": "SmartSchedule API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@smartschedule.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:7070",
    "basePath": "/api/v1",
    "paths": {
...
        "/private/meetings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lấy danh sách sự kiện của người dùng",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meeting"
                ],
                "summary": "Lấy danh sách sự kiện",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.EventResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Tạo một sự kiện hẹn mới với người tham gia",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meeting"
                ],
                "summary": "Tạo sự kiện hẹn",
                "parameters": [
                    {
                        "description": "Thông tin sự kiện",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-api-starter_modules_meeting_dto.CreateEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/private/meetings/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lấy chi tiết một sự kiện hẹn theo ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meeting"
                ],
                "summary": "Lấy thông tin sự kiện",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.EventResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cập nhật thông tin sự kiện hẹn",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meeting"
                ],
                "summary": "Cập nhật sự kiện",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Thông tin cập nhật",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Xóa một sự kiện hẹn",
                "tags": [
                    "Meeting"
                ],
                "summary": "Xóa sự kiện",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/private/meetings/{id}/find-slots": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Tìm các khung giờ phù hợp cho sự kiện",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meeting"
                ],
                "summary": "Tìm khung giờ rảnh",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Tham số tìm kiếm",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.FindSlotsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.FindSlotsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/private/meetings/{id}/select-slot": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Chọn khung giờ cho sự kiện và tạo lịch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meeting"
                ],
                "summary": "Chọn khung giờ",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Thông tin khung giờ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SelectSlotRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.EventPreferences": {
            "type": "object",
            "properties": {
                "exclude_weekends": {
                    "type": "boolean"
                },
                "only_business_hours": {
                    "type": "boolean"
                },
                "prefer_afternoon": {
                    "type": "boolean"
                },
                "prefer_morning": {
                    "type": "boolean"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
        "dto.EventResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration_minutes": {
                    "type": "integer"
                },
                "end_date": {
                    "type": "string"
                },
                "host_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "meeting_link": {
                    "type": "string"
                },
                "participants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ParticipantResponse"
                    }
                },
                "preferences": {
                    "$ref": "#/definitions/dto.EventPreferences"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.FindSlotsRequest": {
            "type": "object",
            "properties": {
                "search_days": {
                    "description": "Alternative: search next N days",
                    "type": "integer"
                },
                "search_end_date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "search_start_date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                }
            }
        },
        "dto.FindSlotsResponse": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "string"
                },
                "participants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ParticipantResponse"
                    }
                },
                "slots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SuggestedSlotDTO"
                    }
                }
            }
        },
...
        "dto.ParticipantResponse": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "string"
                },
                "has_calendar_connected": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.SelectSlotRequest": {
            "type": "object",
            "properties": {
                "end_time": {
                    "description": "RFC3339 format",
                    "type": "string"
                },
                "slot_id": {
                    "type": "string"
                },
                "start_time": {
                    "description": "RFC3339 format",
                    "type": "string"
                }
            }
        },
...
        "dto.SuggestedSlotDTO": {
            "type": "object",
            "properties": {
                "available_count": {
                    "type": "integer"
                },
                "day_of_week": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "formatted_date": {
                    "type": "string"
                },
                "formatted_time": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateEventRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration_minutes": {
                    "type": "integer",
                    "maximum": 480,
                    "minimum": 15
                },
                "preferences": {
                    "$ref": "#/definitions/dto.EventPreferences"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "errors.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/errors.ErrorCode"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "errors.ErrorCode": {
            "type": "integer",
            "enum": [
                1000,
                1001,
                1002,
                1003,
                1004,
                1005,
                2006,
                2007,
                2008,
                2009,
                2010,
                3011,
                3012,
                3013,
                3014,
                3015,
                3016,
                3017,
                3018,
                3019,
                3020,
                3021,
                4022,
                4023,
                4024,
                4025,
                5026,
                5027,
                5028,
                5029,
                6030,
                6031,
                6032,
                6033,
                7034,
                7035
            ],
            "x-enum-varnames": [
                "ErrInvalidCredentials",
                "ErrTokenExpired",
                "ErrUnauthorized",
                "ErrForbidden",
                "ErrMissingAuthorizationHeader",
                "ErrInvalidTokenFormat",
                "ErrInvalidInput",
                "ErrInvalidEmail",
                "ErrInvalidPassword",
                "ErrInvalidFormat",
                "ErrInvalidRequestData",
                "ErrNotFound",
                "ErrAlreadyExists",
                "ErrResourceLocked",
                "ErrResourceExpired",
                "ErrCreateFailed",
                "ErrGetFailed",
                "ErrInvalidSlug",
                "ErrDuplicateSlug",
                "ErrInvalidParent",
                "ErrUpdateFailed",
                "ErrDeleteFailed",
                "ErrDatabase",
                "ErrDatabaseTimeout",
                "ErrUniqueViolation",
                "ErrForeignKey",
                "ErrBusinessRule",
                "ErrInvalidState",
                "ErrLimitExceeded",
                "ErrOperationFailed",
                "ErrInternalServer",
                "ErrConfiguration",
                "ErrThirdParty",
                "ErrNetwork",
                "ErrUploadToR2",
                "ErrSaveStorageDb"
            ]
        },
...
        "go-api-starter_modules_meeting_dto.CreateEventRequest": {
            "type": "object",
            "required": [
                "duration_minutes",
                "title"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration_minutes": {
                    "type": "integer",
                    "maximum": 480,
                    "minimum": 15
                },
                "participants": {
                    "description": "user_ids",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "preferences": {
                    "$ref": "#/definitions/dto.EventPreferences"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}